Refactor run_minimal to a common file in tests
